(#688) follow up
