+++ gamestonk_terminal/stocks/dark_pool_shorts/yahoofinance_view.py
@@ -1,0 +1,41 @@
+__docformat__ = "numpy"
+
+import os
+from tabulate import tabulate
+from governance.helper_funcs import export_data
+from gamestonk_terminal.stocks.dark_pool_shorts import yahoofinance_model
+
+
+def display_most_shorted(num_stocks: int, export: str):
+    """Display most shorted stocks screener. [Source: Yahoo Finance]
+    Parameters
+    ----------
+    num_stocks: int
+        Number of stocks to display
+    export : str
+        Export dataframe data to csv,json,xlsx file
+    """
+    df = yahoofinance_model.get_most_shorted().head(num_stocks)
+    df.dropna(how="all", axis=1, inplace=True)
+    df = df.replace(float("NaN"), "")
+
+    if df.empty:
+        print("No data found.")
+    else:
+        print(
+            tabulate(
+                df,
+                headers=df.columns,
+                floatfmt=".2f",
+                showindex=False,
+                tablefmt="fancy_grid",
+            )
+        )
+    print("")
+
+    export_data(
+        export,
+        os.path.dirname(os.path.abspath(__file__)),
+        "shorted",
+        df,
+    )
